Make periodic fetching work.
This is a really subtle bug.  I tried to reference the current actor in the constructor of Feeder, in order to kick off a 3rd party actor which can communicate with it.

But, because the actor isn't really set up until you call start(), Actor.self was returning an ActorProxy for the current thread (main).  So, sending messages to that ActorProxy would never be acknowledged.

Instead, rely on a more explicit init message being sent.